Rails 3 refactor and gem reorganization
