+++ init.rb
@@ -1,2 +1,1 @@
-require "#{File.dirname(__FILE__)}/lib/active_record/acts/dag"
-ActiveRecord::Base.send(:include, ActiveRecord::Acts::Dag)
+require "#{File.dirname(__FILE__)}/lib/dag"
